feat: add budget type
A type is use to split budget in two categories.
The first one is NORMAL budgets. Budget in this category behave normaly like all simple budgets does. When an expense is added the available amount goes down.
The second categroy is RESERVE budgets. Budget in this category allow users to accumulate available amount over the months.